+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### CHANGES
 
 - Update `cmd/tools/pmanager`: delete constant from crypto/symmetric package, update keySize 32 -> 33 bytes
+- Update `pkg/network/conn`: add context to NewConn function
 
 <!-- ... -->